# Conflicts:
#	src/bp-templates/bp-nouveau/css/buddypress-rtl.min.css
#	src/bp-templates/bp-nouveau/css/buddypress.min.css
#	src/languages/buddyboss.pot

+++ src/bp-moderation/classes/class-bp-moderation.php
@@ -382,7 +382,7 @@ class BP_Moderation {
 
 		// Scope takes precedence.
 		if ( ! empty( $r['filter_query'] ) ) {
-			$filter_query = new BP_Moderation_Query( $r['filter_query'] );
+			$filter_query = new BP_Moderation_Query( $r['filter_query'], $r );
 			$sql          = $filter_query->get_sql();
 
 			if ( ! empty( $sql['where'] ) ) {
@@ -436,6 +436,9 @@ class BP_Moderation {
 
 			$user_ids                    = implode( ',', wp_parse_id_list( $r['user_id'] ) );
 			$where_conditions['user_id'] = "mr.user_id IN ({$user_ids})";
+			if ( ! isset( $r['user_report'] ) ) {
+				$where_conditions['user_id'] .= " and mr.user_report = 0 ";
+			}
 		}
 
 		// Exclude specified items.
@@ -1383,7 +1386,7 @@ class BP_Moderation {
 	public function unhide() {
 		$this->hide_sitewide = 0;
 
-		if ( ! empty( $this->report_id ) ) {
+		if ( ! empty( $this->report_id ) && ! is_admin() ) {
 			$this->delete();
 		} else {
 
@@ -1441,6 +1444,15 @@ class BP_Moderation {
 		if ( 0 === $this->count ) {
 			$wpdb->update( $bp->moderation->table_name, array( 'reported' => 0 ), array( 'id' => $this->id ) ); // phpcs:ignore
 		}
+
+		/**
+		 * Fires after an moderation report item has been deleted to the database.
+		 *
+		 * @since BuddyBoss 2.1.4
+		 *
+		 * @param BP_Moderation $this Current instance of moderation item being deleted. Passed by reference.
+		 */
+		do_action_ref_array( 'bb_moderation_after_delete', array( &$this ) );
 
 		return ! empty( $updated_row );
 	}
